split main into new files
